Speed up AnnotateGlobalSuite.test. Also move it to aggregator suite and rename it appropriately.
