+++ src/test/scala/is/hail/methods/AnnotateGlobalSuite.scala
@@ -5,60 +5,11 @@ import is.hail.annotations.Annotation
 import is.hail.expr.types._
 import is.hail.utils._
 import is.hail.testUtils._
+import org.apache.spark.sql.Row
 import org.apache.spark.util.StatCounter
 import org.testng.annotations.Test
 
 class AnnotateGlobalSuite extends SparkSuite {
-  @Test def test() {
-
-    var vds = hc.importVCF("src/test/resources/sample2.vcf")
-    vds = TestUtils.splitMultiHTS(vds)
-    vds = VariantQC(vds)
-    vds = SampleQC(vds)
-
-    val (afDist, _) = vds.aggregateRows("AGG.map(v => va.qc.AF).stats()")
-    val (singStats, _) = vds.aggregateCols("AGG.filter(sa => sa.qc.n_singleton > 2L).count()")
-    val (acDist, _) = vds.aggregateRows("AGG.map(v => va.qc.AC.toFloat64).stats()")
-    val (crStats, _) = vds.aggregateCols("AGG.map(s => sa.qc.call_rate).stats()")
-
-    val qSingleton = vds.querySA("sa.qc.n_singleton")._2
-
-    val sCount = vds.colValues.value.count(sa =>
-      qSingleton(sa).asInstanceOf[Long] > 2)
-
-    assert(singStats == sCount)
-
-    val qAF = vds.queryVA("va.qc.AF")._2
-    val afSC = vds.variantsAndAnnotations.map(_._2)
-      .aggregate(new StatCounter())({ case (statC, va) =>
-        val af = Option(qAF(va))
-        af.foreach(o => statC.merge(o.asInstanceOf[Double]))
-        statC
-      }, { case (sc1, sc2) => sc1.merge(sc2) })
-
-    assert(afDist == Annotation(afSC.mean, afSC.stdev, afSC.min, afSC.max, afSC.count, afSC.sum))
-
-    val qAC = vds.queryVA("va.qc.AC")._2
-    val acSC = vds.variantsAndAnnotations.map(_._2)
-      .aggregate(new StatCounter())({ case (statC, va) =>
-        val ac = Option(qAC(va))
-        ac.foreach(o => statC.merge(o.asInstanceOf[Int]))
-        statC
-      }, { case (sc1, sc2) => sc1.merge(sc2) })
-
-    assert(acDist == Annotation(acSC.mean, acSC.stdev, acSC.min, acSC.max, acSC.count, acSC.sum))
-
-    val qCR = vds.querySA("sa.qc.call_rate")._2
-    val crSC = vds.colValues.value
-      .aggregate(new StatCounter())({ case (statC, sa) =>
-        val cr = Option(qCR(sa))
-        cr.foreach(o => statC.merge(o.asInstanceOf[Double]))
-        statC
-      }, { case (sc1, sc2) => sc1.merge(sc2) })
-
-    assert(crStats == Annotation(crSC.mean, crSC.stdev, crSC.min, crSC.max, crSC.count, crSC.sum))
-  }
-
   @Test def testTable() {
     val out1 = tmpDir.createTempFile("file1", ".txt")
